Adding the transcriptic launch command to configure and launch protocols
Summary:
transcriptic launch allow you to configure and run a protocol using the web UI.
This can be done for protocols which have not yet been packaged and uploaded. The
result of the input selected will be passed as input to your protocol, and as a
result AutoProtocol will be generated as the final output, suitable for piping to
transcriptic analyze for example.

Example uses:
* transcriptic launch SampleProtocol -p p18rkyfdxyap9 | tee autoprotocol.json
* transcriptic launch SampleProtocol -p p18rkyfdxyap9 | transcriptic analyze

Additional changes included
* Extracted load_manifest and load_protocol into reuseable functions
* Added a run_protocol helper method used by preview and launch
- Added a stdchannel_redirected context manager to temporarily suppress a standard channel
  such stderr or stdout

Test Plan: None

Reviewers: yangchoo

Reviewed By: yangchoo

Subscribers: peter

Maniphest Tasks: T4741

Differential Revision: https://work.r23s.net/D4170