fixed data overwrite issue in _PlateRead
Summary:
When using the _PlateRead class, it is the case that if there are
multiple instances of the same label in group_labels (i.e. if multiple groups
in a plate have the same volume/label), read data for all but the last group
corresponding to a label is overwritten, because in an intermediate step
the data is stored in a dict with labels as keys.

Fixed this by removing the intermediate mapping step, and creating read data
DataFrames directly.

The resulting class is backwards compatible, since the resultant DF behaves
the same as it did before the change.

Test Plan:
Tested locally with several QC runs
Release Test: None

Reviewers: yangchoo, peter

Reviewed By: peter

Differential Revision: https://work.r23s.net/D7288