Conditional display of views based on` feature_groups`
Summary:
This diff controls what's presented to the user based on the `feature_groups` they have access to.
The set of views presented are as discussed in T9662.

Note: There's a little grossness here in terms of extending Click base-classes, but some of that will go away once Click 7 is released

Test Plan: Tested by setting `feature_groups` manually in a local dev environment and making sure that the expected responses are returned when calling `transcriptic`, `transcriptic --help`

Reviewers: eric

Subscribers: alex, evan

Differential Revision: https://work.r23s.net/D7398