Add a `generate_protocol` CLI command that inits an example protocol.
Summary:
Creating a python protocol can be tough going.  One must understand
autoprotocol-py, the format of the manifest.json, and know which
dependencies are required.

The `generate_protocol` command should make it easier for both new and
old users to get up and running when writing protocols.

Test Plan:
Manually attempt to generate a protocol, and verify that it preview succeeds.

  `transcriptic generate_protocol foo`
  `transcriptic preview foo`

Reviewers: yangchoo, peter

Differential Revision: https://work.r23s.net/D9168