docs: Use readthedocs config file
We had a separate doc/requirements.txt file that was previously used for
maintaing documentation dependencies for readthedocs.org

Switching to the readthedocs config file would allow us to remove that
duplicated file and use the right build environment for generating docs.

Also adding in a badge so it'll be more obvious if we break the build in
the future.