feat: Support Bearer Token Authentication (#165)
We will soon be passing a client_credentials (service-to-service) bearer token to programs when executing in the context of Igor. Web will validate the bearer token (and only the bearer token, no user token needed) when receiving an API request.

Programs internally call the transcriptic library (this repo) to form an API Connection.

The change in this PR involves changing the Connection instantiation routine to recognize if the token is a Bearer token and, if so, set the standard OAuth Authorization header (and no X-User-Token header). This is opposed to a non-bearer token, which will result in the usual X-User-Token header.

Jira Issue: SEA-271

Co-authored-by: Yang <yang.choo@strateos.com>