feat: Seed commands tests, rework projects internals
The focus of this diff is to seed example tests using the responses
library for the commands.py file, which currently has no test coverage.
The `projects` and `submit` command, which are among the more utilized
functions are used here as examples.

In the process of adding tests, the projects internals in commands.py
was found to be confusing and made to be more intuitive.
A deprecation warning is attached to the `-i` option in the `projects`
command. Instead of requiring an input which is unused, this turns out
into a flag and renamed to `--names`.