feature: handle new frontend deployment for frontend (#204)
Summary:
The frontend url (lab.core.strateos.com/haven/wctest) is not the same that the url to the frontend node. This would be confusing to user to have to use different one.

The exec method will request the frontend server for the IP of the frontend node, and then make the request to it.

```
$ transcriptic preview Read | transcriptic exec --api http://lab.core.strateos.com/haven/wctest
Sending request to 13.52.223.111:9000
Success. View http://lab.core.strateos.com/haven/wctest/dashboard to see the scheduling outcome.
```